feat(Borders in table cell): :sparkles: Allow removing corners connections from borders
In a table cell, when there are adjacent cells with borders applied in the same position as other cells, a connection is drawn between them as standard behavior. This also occurs for double borders, but it occurs even when there are no adjacent borders, turning the double border into a rectangle.  An option will be inserted in the settings to include or not this connection between the borders.  As a result, even if they are in the same position, cells with adjacent borders will have a small gap between them.

#1